Change project's organization, add a lot of gcc flags and correct new errors it led to.

+++ src/circle.c
@@ -8,33 +8,33 @@ void circle_draw(Canvas* canvas, Circle* circle) {
     Point min;
     Point max;
     Pixel pix;
-
+    
     pix.color = circle->color;
-
+    
     min.x = circle->center.x - circle->radius;
     max.x = circle->center.x + circle->radius;
     min.y = circle->center.y - circle->radius;
     max.y = circle->center.y + circle->radius;
-
+    
     for (i = min.x; i <= max.x; i++) {
-        dx = i - circle->center.x;
-        dy = sqrt(circle->radius * circle->radius - dx * dx);
+        dx = (float) (i - circle->center.x);
+        dy = (float) sqrt((float) (circle->radius * circle->radius) - dx * dx);
         pix.position.x = i;
-
-        pix.position.y = circle->center.y + dy;
+        
+        pix.position.y = (int) ((float) circle->center.y + dy);
         pixel_draw(canvas, &pix);
-        pix.position.y = circle->center.y - dy;
+        pix.position.y = (int) ((float) circle->center.y - dy);
         pixel_draw(canvas, &pix);
     }
-
+    
     for (i = min.y; i <= max.y; i++) {
-        dy = i - circle->center.y;
-        dx = sqrt(circle->radius * circle->radius - dy * dy);
+        dy = (float) (i - circle->center.y);
+        dx = (float) sqrt((float) (circle->radius * circle->radius) - dy * dy);
         pix.position.y = i;
-
-        pix.position.x = circle->center.x + dx;
+        
+        pix.position.x = (int) ((float) circle->center.x + dx);
         pixel_draw(canvas, &pix);
-        pix.position.x = circle->center.x - dx;
+        pix.position.x = (int) ((float) circle->center.x - dx);
         pixel_draw(canvas, &pix);
     }
 }
@@ -46,22 +46,22 @@ void circle_draw_fill(Canvas* canvas, Circle* circle) {
     Point min;
     Point max;
     Pixel pix;
-
+    
     pix.color = circle->color;
-
+    
     min.x = circle->center.x - circle->radius;
     max.x = circle->center.x + circle->radius;
     min.y = circle->center.y - circle->radius;
     max.y = circle->center.y + circle->radius;
-
+    
     for (pix.position.x = min.x; pix.position.x <= max.x; pix.position.x++) {
-
-        dx = pix.position.x - circle->center.x;
-
+    
+        dx = (float) (pix.position.x - circle->center.x);
+        
         for (pix.position.y = min.y; pix.position.y <= max.y; pix.position.y++) {
-
-            dy = pix.position.y - circle->center.y;
-
+        
+            dy = (float) (pix.position.y - circle->center.y);
+            
             if (dx * dx + dy * dy <= circle->radius * circle->radius) {
                 pixel_draw(canvas, &pix);
             }
